Debian 10/buster support
The formula needs to be updated to support Debian 10/buster so that
Wallabag can be installed on the latest release.

It isn't currently possible to initialise a new Wallabag v2.3.8 instance
on Debian 10/buster due to the default PostgreSQL version, but importing
the database from a previous install should still work. See:

    https://github.com/wallabag/wallabag/issues/3739